+++ src/internal/types/combinations.d.ts
@@ -5,12 +5,12 @@ import { InRange8 } from './utility';
 type CombinationsPermutationsByIterable<Iter extends Iterable<any>> = Iter extends Iterable<infer T>
   ? Iter extends T[]
     ? CombinationsPermutationsByLength<T, Iter['length']>
-    : (Iterable<T[]> & { getSize: () => number })
+    : (Iterable<T[]> & { size: number })
   : never;
 
 type CombinationsPermutationsByLength<T, R extends number> = Iterable<
   R extends InRange8 ? Repeat<T, R> : T[]
-> & { readonly getSize: () => number };
+> & { readonly size: number };
 
 /**
  * Function signature of `permutations` and `combinations`