+++ src/methods/combinations-with-replacement/combinations-with-replacement.mjs
@@ -16,21 +16,25 @@ function isSorted(arr) {
 }
 
 export function combinationsWithReplacement(iterable, r) {
-  const arr = Array.from(iterable);
+  const arr = [...iterable];
   const len = arr.length;
+  let _size = null;
+
   r = r === undefined ? len : r;
+
   return {
     *[Symbol.iterator]() {
       const toIndex = i => arr[i];
 
       for (let indices of product(...map(range(0, r), () => range(0, len)))) {
         if (isSorted(indices)) {
-          yield Array.from(map(indices, toIndex));
+          yield [...map(indices, toIndex)];
         }
       }
     },
-    getSize() {
-      return combinationsWithReplacementSize(len, r);
+
+    get size() {
+      return _size === null ? (_size = combinationsWithReplacementSize(len, r)) : _size;
     },
   };
 }