+++ src/methods/combinations/__tests__/combinations.spec.ts
@@ -10,12 +10,12 @@ import { combinations } from '../../..';
 describe('combinations', () => {
   it('returns empty', () => {
     const iter = combinations([]);
-    expect(iter.getSize()).toEqual(0);
+    expect(iter.size).toEqual(0);
     expect(Array.from(iter)).toEqual([]);
   });
   it('returns combinations', () => {
     const iter = combinations([1, 2]);
-    expect(iter.getSize()).toEqual(1);
+    expect(iter.size).toEqual(1);
     expect(Array.from(iter)).toEqual([[1, 2]]);
   });
   it('can be reused', () => {
@@ -25,13 +25,13 @@ describe('combinations', () => {
   });
   it('returns combinations (max n)', () => {
     const iter = combinations(2, [1, 2, 3, 4]);
-    expect(iter.getSize()).toEqual(6);
+    expect(iter.size).toEqual(6);
     const expected = [[1, 2], [1, 3], [1, 4], [2, 3], [2, 4], [3, 4]];
     expect(Array.from(iter)).toEqual(expected);
   });
   it('returns combinations 0', () => {
     const iter = combinations(0, [1, 2, 3, 4]);
-    expect(iter.getSize()).toEqual(0);
+    expect(iter.size).toEqual(0);
     expect(Array.from(iter)).toEqual([]);
   });
 });