+++ src/methods/permutations/__tests__/permutations.spec.ts
@@ -10,12 +10,12 @@ import { permutations } from '../../..';
 describe('permutations', () => {
   it('returns empty', () => {
     const iter = permutations([]);
-    expect(iter.getSize()).toEqual(0);
+    expect(iter.size).toEqual(0);
     expect(Array.from(iter)).toEqual([]);
   });
   it('returns permutations', () => {
     const iter = permutations(2, [1, 2]);
-    expect(iter.getSize()).toEqual(2);
+    expect(iter.size).toEqual(2);
     expect(Array.from(iter)).toEqual([[1, 2], [2, 1]]);
   });
   it('can be reused', () => {
@@ -25,7 +25,7 @@ describe('permutations', () => {
   });
   it('returns permutations', () => {
     const iter = permutations(2, [1, 2, 3, 4]);
-    expect(iter.getSize()).toEqual(12);
+    expect(iter.size).toEqual(12);
     const expected = [
       [1, 2],
       [1, 3],